Work on connections

+++ LookUp/Graph/GraphViewController.swift
@@ -22,7 +22,7 @@ struct GraphViewControllerRepresentable: UIViewControllerRepresentable {
 class GraphViewController: UIViewController {
     var graphViewModel: GraphViewModel
 
-    var spacing = Double(80)
+    var spacing = Double(160)
 
     var containerView = UIView()
